Refactor and add the direct method (we are all sick&tired from the massive changes of qiskit, etc.)
